add: split check_and_inference func function
