+++ src/type_parser/check_and_inference/func/mod.rs
@@ -1,0 +1,1 @@
+pub mod func;
