+++ src/type_parser/check_and_inference/mod.rs
@@ -2,3 +2,4 @@ pub mod check_and_inference;
 mod _struct;
 mod variable_declaration;
 mod expr_stmt;
+mod func;
